MediaLiveClientのconstructorや複雑な処理をHelperに集約
